fix: replace real requests with mock requests
The requests were still being sent to the real endpoint, outside the device. The problem was that replayActive in driverCommunicationService was not working as expected, at least that's what I understood. I changed the condition using part of the session mate backend url to avoid using a mock on any session mate request.